Hooking up some of the logic with the dialog
